Decompose server application wiring
